# docs/openapi.json
{"openapi": "3.1.0", "info": {"title": "RESTai", "description": "RESTai is an AIaaS (AI as a Service) platform. Create AI projects and consume them via REST API.\n\nSupports multiple project types: **RAG**, **Inference**, **Agent**, **Vision**, **Router**, and **RAG-SQL**.\n\n## Authentication\n\nAll endpoints require authentication via one of:\n- **JWT Cookie** (`restai_token`)\n- **Bearer API Key** (`Authorization: Bearer <key>`)\n- **Basic Auth** (`Authorization: Basic <credentials>`)\n", "version": "6.0.0"}, "paths": {"/oauth/{provider}/login": {"get": {"tags": ["Auth"], "summary": "Oauth Login", "description": "Initiate OAuth login flow for the specified provider.", "operationId": "oauth_login_oauth__provider__login_get", "parameters": [{"name": "provider", "in": "path", "required": true, "schema": {"type": "string", "description": "OAuth provider name", "title": "Provider"}, "description": "OAuth provider name"}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/oauth/{provider}/callback": {"get": {"tags": ["Auth"], "summary": "Oauth Callback", "description": "Handle OAuth callback from the specified provider.", "operationId": "oauth_callback_oauth__provider__callback_get", "parameters": [{"name": "provider", "in": "path", "required": true, "schema": {"type": "string", "description": "OAuth provider name", "title": "Provider"}, "description": "OAuth provider name"}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}}, "components": {"schemas": {"HTTPValidationError": {"properties": {"detail": {"items": {"$ref": "#/components/schemas/ValidationError"}, "type": "array", "title": "Detail"}}, "type": "object", "title": "HTTPValidationError"}, "ValidationError": {"properties": {"loc": {"items": {"anyOf": [{"type": "string"}, {"type": "integer"}]}, "type": "array", "title": "Location"}, "msg": {"type": "string", "title": "Message"}, "type": {"type": "string", "title": "Error Type"}, "input": {"title": "Input"}, "ctx": {"type": "object", "title": "Context"}}, "type": "object", "required": ["loc", "msg", "type"], "title": "ValidationError"}}}}